[Step4] 코드리뷰 반영 - UI로직과 핵심로직 분리 (resultView)
