+++ src/main/kotlin/racing/race/CarRace.kt
@@ -5,14 +5,18 @@ import racing.common.CarRaceProperty
 import racing.utils.RandomNumberGenerator.getRandomNumber
 import racing.view.ResultView
 
-class CarRace(private val carRaceProperty: CarRaceProperty) {
-    fun start(resultView: ResultView) {
+class CarRace(
+    private val carRaceProperty: CarRaceProperty,
+    private val resultView: ResultView
+) {
+    fun start() {
         for (round in CarRaceProperty.ROUND_START..carRaceProperty.roundSize) {
             raceByRound()
             resultView.printResultByRound(round)
         }
 
-        resultView.printWinner()
+        val winnerJudge = WinnerJudge(carRaceProperty)
+        resultView.printWinner(winnerJudge.getWinnerCarNames())
     }
 
     private fun raceByRound() {
