+++ src/test/kotlin/racing/CarRaceTest.kt
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import racing.common.CarRaceProperty
 import racing.race.CarRace
+import racing.view.ResultView
 
 class CarRaceTest {
 
@@ -12,8 +13,9 @@ class CarRaceTest {
         val roundSize = 1
         val carNames = listOf("A")
         val carRaceProperty = CarRaceProperty(roundSize, carNames)
+        val resultView = ResultView(carRaceProperty)
 
-        val carRace = CarRace(carRaceProperty)
+        val carRace = CarRace(carRaceProperty, resultView)
         carRace.moveCarPosition(carRaceProperty.cars[0])
 
         assertThat(carRaceProperty.cars[0].getMoveCount()).isEqualTo(1)
